PLANNER-580 remove Helper classes + bring enum switches into Configs that own them
