+++ optaplanner-core/src/main/java/org/optaplanner/core/config/heuristic/selector/entity/EntitySelectorConfig.java
@@ -23,6 +23,7 @@ import java.util.List;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamAsAttribute;
 import com.thoughtworks.xstream.annotations.XStreamImplicit;
+import org.optaplanner.core.api.domain.entity.PlanningEntity;
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.config.heuristic.policy.HeuristicConfigPolicy;
 import org.optaplanner.core.config.heuristic.selector.SelectorConfig;
@@ -427,10 +428,10 @@ public class EntitySelectorConfig extends SelectorConfig<EntitySelectorConfig> {
             SelectionSorter sorter;
             if (sorterManner != null) {
                 EntityDescriptor entityDescriptor = entitySelector.getEntityDescriptor();
-                if (!EntitySorterMannerHelper.hasSorter(sorterManner, entityDescriptor)) {
+                if (!hasSorter(sorterManner, entityDescriptor)) {
                     return entitySelector;
                 }
-                sorter = EntitySorterMannerHelper.determineSorter(sorterManner, entityDescriptor);
+                sorter = determineSorter(sorterManner, entityDescriptor);
             } else if (sorterComparatorClass != null) {
                 Comparator<Object> sorterComparator = ConfigUtils.newInstance(this,
                         "sorterComparatorClass", sorterComparatorClass);
@@ -565,6 +566,41 @@ public class EntitySelectorConfig extends SelectorConfig<EntitySelectorConfig> {
     @Override
     public String toString() {
         return getClass().getSimpleName() + "(" + entityClass + ")";
+    }
+
+    public static boolean hasSorter(EntitySorterManner entitySorterManner, EntityDescriptor entityDescriptor) {
+        switch (entitySorterManner) {
+            case NONE:
+                return false;
+            case DECREASING_DIFFICULTY:
+                return true;
+            case DECREASING_DIFFICULTY_IF_AVAILABLE:
+                return entityDescriptor.getDecreasingDifficultySorter() != null;
+            default:
+                throw new IllegalStateException("The sorterManner ("
+                        + entitySorterManner + ") is not implemented.");
+        }
+    }
+
+    public static SelectionSorter determineSorter(EntitySorterManner entitySorterManner, EntityDescriptor entityDescriptor) {
+        SelectionSorter sorter;
+        switch (entitySorterManner) {
+            case NONE:
+                throw new IllegalStateException("Impossible state: hasSorter() should have returned null.");
+            case DECREASING_DIFFICULTY:
+            case DECREASING_DIFFICULTY_IF_AVAILABLE:
+                sorter = entityDescriptor.getDecreasingDifficultySorter();
+                if (sorter == null) {
+                    throw new IllegalArgumentException("The sorterManner (" + entitySorterManner
+                            + ") on entity class (" + entityDescriptor.getEntityClass()
+                            + ") fails because that entity class's @" + PlanningEntity.class.getSimpleName()
+                            + " annotation does not declare any difficulty comparison.");
+                }
+                return sorter;
+            default:
+                throw new IllegalStateException("The sorterManner ("
+                        + entitySorterManner + ") is not implemented.");
+        }
     }
 
 }
